Default to index.md in subdirectories.
If an access to '#!/test/' is made, we will use the trailing slash as an
indication that 'test' is a directory. We will try to load
'test/index.md'. This lets create organized subdirectories with index.md files
to create pretty URLs instead of having to link to #!/test/index explicitly.